UI logic decoupled from controller logic - tests passing
